Build changes and testing
The build setup for the project has been corrected to allow for the linking of the static library.
Testing has been added to the project in the form of the test directory. A .c smoke test has been
added. The library init function has been reworked to prevent seg fault, but is still not working.